operating systems
a multithreaded program in C that outputs prime numbers. This program work as follows: 
The user will run the program and will enter a number on the commandline. 
The program will then create a separate thread that outputs all the prime numbers less than or equal to the number entered by the user.